Converted to simpler driver
As promised this commit reduces the complexity of the driver by using
another led driver.  The interface uses sysfs files for interaction
and the idea is to simply echo 1 into the right file to make a
led light the right color.

Right now all we want to do is to see if it finds the device.