Working driver
This driver now works with the Nelson's firmware.  You can now write
to the /sys/bus/usb/drivers/teensi/2-1:1/{blue,green,red}.  It will
work fine but there are problems with error -110 which I believe is
a time out error.  It could be a firmware issue, but an equivalent
libusb driver seems to work.